RANGER-4157: fixed handling of implicit addition of expression condition
(cherry picked from commit e89016cc581379b5c98f99de202c967f503dcfe3)
